+++ sdk/resourcemanager/hybridcompute/armhybridcompute/client_factory.go
@@ -44,6 +44,14 @@ func (c *ClientFactory) NewExtensionMetadataClient() *ExtensionMetadataClient {
 	}
 }
 
+// NewGatewaysClient creates a new instance of GatewaysClient.
+func (c *ClientFactory) NewGatewaysClient() *GatewaysClient {
+	return &GatewaysClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
+}
+
 // NewMachineExtensionsClient creates a new instance of MachineExtensionsClient.
 func (c *ClientFactory) NewMachineExtensionsClient() *MachineExtensionsClient {
 	return &MachineExtensionsClient{
@@ -110,6 +118,14 @@ func (c *ClientFactory) NewPrivateLinkResourcesClient() *PrivateLinkResourcesCli
 // NewPrivateLinkScopesClient creates a new instance of PrivateLinkScopesClient.
 func (c *ClientFactory) NewPrivateLinkScopesClient() *PrivateLinkScopesClient {
 	return &PrivateLinkScopesClient{
+		subscriptionID: c.subscriptionID,
+		internal:       c.internal,
+	}
+}
+
+// NewSettingsClient creates a new instance of SettingsClient.
+func (c *ClientFactory) NewSettingsClient() *SettingsClient {
+	return &SettingsClient{
 		subscriptionID: c.subscriptionID,
 		internal:       c.internal,
 	}