+++ sdk/resourcemanager/hybridcompute/armhybridcompute/options.go
@@ -18,6 +18,40 @@ type ExtensionMetadataClientListOptions struct {
 	// placeholder for future optional parameters
 }
 
+// GatewaysClientBeginCreateOrUpdateOptions contains the optional parameters for the GatewaysClient.BeginCreateOrUpdate method.
+type GatewaysClientBeginCreateOrUpdateOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// GatewaysClientBeginDeleteOptions contains the optional parameters for the GatewaysClient.BeginDelete method.
+type GatewaysClientBeginDeleteOptions struct {
+	// Resumes the LRO from the provided token.
+	ResumeToken string
+}
+
+// GatewaysClientGetOptions contains the optional parameters for the GatewaysClient.Get method.
+type GatewaysClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// GatewaysClientListByResourceGroupOptions contains the optional parameters for the GatewaysClient.NewListByResourceGroupPager
+// method.
+type GatewaysClientListByResourceGroupOptions struct {
+	// placeholder for future optional parameters
+}
+
+// GatewaysClientListBySubscriptionOptions contains the optional parameters for the GatewaysClient.NewListBySubscriptionPager
+// method.
+type GatewaysClientListBySubscriptionOptions struct {
+	// placeholder for future optional parameters
+}
+
+// GatewaysClientUpdateOptions contains the optional parameters for the GatewaysClient.Update method.
+type GatewaysClientUpdateOptions struct {
+	// placeholder for future optional parameters
+}
+
 // MachineExtensionsClientBeginCreateOrUpdateOptions contains the optional parameters for the MachineExtensionsClient.BeginCreateOrUpdate
 // method.
 type MachineExtensionsClientBeginCreateOrUpdateOptions struct {
@@ -213,5 +247,20 @@ type PrivateLinkScopesClientListOptions struct {
 
 // PrivateLinkScopesClientUpdateTagsOptions contains the optional parameters for the PrivateLinkScopesClient.UpdateTags method.
 type PrivateLinkScopesClientUpdateTagsOptions struct {
+	// placeholder for future optional parameters
+}
+
+// SettingsClientGetOptions contains the optional parameters for the SettingsClient.Get method.
+type SettingsClientGetOptions struct {
+	// placeholder for future optional parameters
+}
+
+// SettingsClientPatchOptions contains the optional parameters for the SettingsClient.Patch method.
+type SettingsClientPatchOptions struct {
+	// placeholder for future optional parameters
+}
+
+// SettingsClientUpdateOptions contains the optional parameters for the SettingsClient.Update method.
+type SettingsClientUpdateOptions struct {
 	// placeholder for future optional parameters
 }